valve_on race condition fixed
cleaned up debug output

changed backyard zones to 5 minutes

removed some 24V relay code that isn't needed (we keep 24V on all the time)

created valve_on() function to check if a valve is on instead of using the valve_on variable.  There was a race condition between MQTT messaging and the valve 'state machine' where an MQTT message could clear the valve_on variable if a valve-off message was sent for a valve that wasn't on.  The valve_on() function looks at the GPIO pin states to determine on/off now.